+++ Lunar/More Communities Views/ExploreCommunitiesView.swift
@@ -42,7 +42,7 @@ struct ExploreCommunitiesView: View {
               communityIcon: community.community.icon
             )
           } label: {
-            CommunityRowView(community: convertToRealmCommunity(community: community))
+            CommunityItem(community: convertToRealmCommunity(community: community))
               .environmentObject(communitiesFetcher)
           }
         }